RL-Highway-Env-Project/requirements-py310.txt
absl-py==0.15.0
astunparse==1.6.3
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
clang==5.0
cloudpickle==2.0.0
cycler==0.10.0
decorator==5.1.0
dm-tree==0.1.6
flatbuffers==1.12
gast==0.4.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
h5py==3.7.0
gym==0.26.5
numpy==1.23.5
highway-env==1.4
idna==3.3
Keras-Preprocessing==1.1.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.5.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-probability==0.16.0
Pillow==8.4.0
protobuf==3.19.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
termcolor==1.1.0
tqdm==4.62.3
typing-extensions==3.7.4.3
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.12.1
